Extract notification events to its own component
